add data transfer between 2 devices
the data transfer between 2 devices is completed, the data will be saved in the database and will be updated in the ui accordingly.
TODO: update the ui in the sync screen to show the user something is happening ;)

+++ fittrackapp/src/main/java/com/berbas/fittrackapp/screens/profile/ProfileScreen.kt
@@ -82,7 +82,7 @@ fun ProfileScreen(
                         onWeightChange = { newValue ->
                             profileViewModel.onEvent(
                                 PersonEvent.SetWeight(
-                                    newValue.toDouble()
+                                    newValue
                                 )
                             )
                         },
@@ -90,7 +90,7 @@ fun ProfileScreen(
                         onHeightChange = { newValue ->
                             profileViewModel.onEvent(
                                 PersonEvent.SetHeight(
-                                    newValue.toDouble()
+                                    newValue
                                 )
                             )
                         }
